--- request_fulfillment_authority/Dispatcher/configuration/db.config
@@ -1,5 +1,0 @@
-[database]
-user=root
-pass=root
-host=localhost
-schema=AUTONOMOUS_DRIVING_ASSISTANCE
